+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Next Release
 -
 
+v1.2.0
+* Tracked image snap now emailing
+* Added threading for improved performance
+
 v1.1.0
 * Camera tracking logic implemented
 * Added email service